[FEATURE] Automatically create a useable processing folder
When the file storage record is saved, a DataHandlerHook will check that the directory is not on a Google Drive™, and that it is public and writeable.

If the processing folder does not pass this test or is an empty string it will be set to "0:typo3temp/assets/_processed_storage_[UID]" where [UID] is the UID of the file storage record.